requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==3.2.10
django-cors-headers==4.3.1
django-filter==2.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.7
Markdown==3.3.4
mssql-django==1.5
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
pyodbc==5.1.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2021.1
requests==2.32.3
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
